Add support for build containers
For every image, the presence of a Dockerfile.build.template will be
checked for, and if that's present, it will be used to produce artifacts
that should reside inside the /build directory of the build container,
and will be extracted to the build subdirectory.

Also add a new filter for pruning packages we no longer need.

Change-Id: I2901aec8aae886c0347eee9669e2072cdc3983f6