Create an envoy docker image.
There is no convenient way to reuse the dockerfile that envoproxy offers in
our current build system and with our currently supported distros (the
official image uses ubuntu xenial). So we merge together the script that
is used to generate the "build image" and the steps used to run it and
actually build the software.

Change-Id: I4107dab6ca361ed20a8e0daaac02644eb28d74a9