Add an image for python2 app based on Buster
Zuul runs on Jessie and we need to upgrade the host to Buster. It is a
python 2 application and we need a build image. Our time frame is too
limited to attempt completing the python 3.

Add a PYTHON_SUFFIX variable to the common run.sh entry point to avoid
copy pasting it. It defaults to '3' and is set to '2' for this image.

Bug: T215458
Change-Id: Ie6d5988b2404ebb175a7ebdb224516934b98c88d