Switch the base image to buster from stretch.
As a consequence, the following images will be rebuilt with a buster
base:

* coredns
* nutcracker
* prometheus-statsd-exporter
* service-checker
* wmfdebug

these are all images that are final-use images so are deemed a lesser
risk. we can probably do the same with python and ruby, but that will
require to check the users of these images.

Also fix the naming of the golang-stretch image so that it doesn't
collide with the other golang image.

Change-Id: I68ad8c3db50593760878e03f1922479bcdbfd2f6