ratelimit: crash on startup if config is invalid
Ratelimiter service silently ignores invalid config.
The service supports reloading config on the fly by watching
the config dir, and it makes sense that it doesn't want to crash
if it's already working. However, on startup we would prefer
different behaviour. The service provides a config_check
utility for this purpose, so we use it.

I've had to change the base image to debian, since scratch doesn't
have a shell, and copying it over seems like too much overhead.
Easier to use debian.

Change-Id: I80bed1b3459f03e5564fd98d862bed1e8f81183c