Add apache httpd base image
This image will be the base for our MediaWiki installation. Please
note this image is still not solving any of the basic logging issues
explained at T265876 because it's supposed to be our base image.

Also: in this first version, we're not optimizing for size, but
the image is currently 178 MB, and can probably be reduced by a lot.

Bug: T265324
Change-Id: Ia90ace1667099c45bb7d3ff60dbd113438b36d63