php: make env variables play nice with blubber
The env variables we've used for the php images were (on purpose) not
bash compatible, but would still work both in docker and in php.inis.

When I tried to modify one in a blubber.yaml file, that resulted in a
validation error, as blubber validates that env variables are strictly
POSIX compliant.

While I think it should validate them according to docker's logic, it's
better to adapt the images for now.

Also, allow disabling opcache for development purposes.

Change-Id: Id2c7fbd6cf381eb7e23981c69c47c3e02225bda1