Add httpd image for MediaWiki
This image is based on the httpd-fcgi one, and adds the configuration
that will allow to run MediaWiki in production, as well in a local dev
environment. It doesn't include any actual virtualhosts, as those will be
injected at the higher layers according to the specific installation.

Specifically, this image does the following:
* Disable mod_deflate, and enable mod_access_compat
* Inject a specific apache2 configuration
* Inject our configuration for autoindex, expire and mime modules
* Overwrite the default virtualhosts
* Add a catchall for non-configured domains
* Inject the SERVERGROUP variable in backend calls

Please note that some of this is duplication of what is already in puppet.
I considered these files to change seldom enough that taking the hit makes
sense.

Change-Id: I04d22fecb48ed13689df33516ceeec6e74458ce6