istio: add a more specific build target
Without this patch, the build image that gets pulled
is the latest published in the registry, not the latest
in the build's changelog. After the last run of the image
build script, we have:

- proxyv2 failed
- operator and pilot succeded (but binaries were copied
  from the wrong build)
- build published correctly

This change should allow to build the 1.9.5-1 version of
proxyv2, and the 1.9.5-2 version of pilot and operator.

Bug: T278192
Change-Id: I81f23781fd11a1e61073fa3d1c63577bc08a760d