kubeflow: add workaround for TLS validation in storage-initializer
This workaround is not great, but it seems the only compromise to
allow boto to contact thanos-swift (our s3 endpoint until MOSS is
up and running) without erroring out. The storage-initializer
container is not listed in a traditional k8s container spec, but
as part of a config-map, without the possibility of adding ENV
variables. For more info
https://github.com/kubeflow/kfserving/blob/release-0.6/config/configmap/inferenceservice.yaml#L138-L145
We'll likely need to follow up with upsteam, but in the meantime
this change could allow us to proceed with testing.

Bug: T272919
Change-Id: Ib37ee3282207edb83b3f796bdef74eac92d03766