knative: backport patch to tune pod DNS settings from version 1.5
We are currently observing a constant flow of DNS queries to
ml-serve's coredns, mostly due to some Istio/Envoy peculiarities
in service discovery (see https://github.com/knative/serving/issues/12917).

One of the workaround suggested is to tune the pods' resolv.con file,
setting ndots to a value of 2/3. This is currently supported by
KServe but not by Knative, so this patch attempts to backport
the feature from version 1.5.

The upstream patch didn't apply cleanly so I had to adapt it
a little for our version, 0.18.1, but it is simple enough
that it should work without requiring extra code/dependencies
etc..

Bug: T313915
Change-Id: Ib99e02d67e2c72f63dd9b22f15b0093d0e22960e