Add a new production images for spark and spark-operator
This image generation pipeline takes inspiration from the Dockerfiles
included with the spark and spark-operator sources.Three new images are
produced and published to the registry in this order.

* spark-build
This is the build environment for both the spark distribution and the
spark operator.
* spark
This copies the binary spark distribution from spark-build and
configures an execution environment for it.
* spark-operator
This builds from the spark image and adds the spark-operator binary from
spark-build along with a new entrypoint.

Bindings for Python and R are included and we have selected Hadoop
version 2.10.2

Since this is a maven build we need to add support for the webproxy to
the build options in production.

Bug: T318730
Change-Id: I17fa8806fac21df005949e574d2373dbc2890d3d