istio: reduce Envoy logspam
The Kafka logging cluster is periodically spammed by the ml-serve
clusters, especially during deployments. One of the biggest
source of spam is Envoy, due to:

https://github.com/istio/istio/pull/32387

The above patch applies (almost) cleanly to the 1.9.5 branch
and it seems easy enough to be backported now, so solve
the logspam issue even before any migration to a higher version
of Istio.

The upstream 1.9.5-patch branch doesn't exist anymore,
so the build is switched to the 1.9.5 git tag plus
https://github.com/istio/istio/pull/32792.

Bug: T320468
Change-Id: I94cdedb0e5e8164a390f7a049c28bf1afbd523e4