coredns: upgrade to 1.8.7
We'd need to update coredns to 1.8.x for the upcoming migration
to k8s 1.23. The current Docker images uses a dedicated
Debian repository for coredns, without using a multi-stage
build process as we do for most of our images.

This change moves to a multi-stage build process, and updates
the coredns image's go code to 1.8.7.

Bug: T321159
Change-Id: I5fb8cfd2fc2ff36208048e6b58b10608bc554870