Update the spark and spark-operator images
This is a second iteration of the spark and spark-operator images, which
were required in order to address an issue with launching the executors.

Several of the techniques employed by the upstream authors were not
desirable in our environment, so we have decided to supply our own
versions of the entrypoint scripts.

One additional script that generates certificates is also included,
althought this functionality may be replaced with cfssl-issuer in the
near future.

Since this script uses curl to communicate with the kubernetes API, we
also need to install the wmf-certificates package to the operator image.

Bug: T318730
Change-Id: Idf5e337e8809694e4000ac08a0b0948b80b4086b