Update the spark images to remove upstream support for the webhook
We previoulsy updated the images to include a script provided by the
upstream project called 'gencerts.sh'. This script had two functions:

1: Create an x509 certificate and private key to secure webhook traffic
2: Create an auth token in order to authenticate webhook clients

It used calls to the K8S API in order to share these secrets between the
operator and the spark applications.

However, we have decided that for this phase of testing we will not be
using the webhook. When we do want to test, we will use our existing
mechanism for handling secrets and x509 keypairs, rather than the
methods employed by the upstream authors.

Also, several suggestions for tidying the entrypoint.sh script for the
spark image and simplifying the proxy selectionwere incorporated into
this change.

Bug: T318926
Change-Id: I65569d04065dde3aca5d2a80ffa9ae9e28115a08