Downgrade the version of spark to 3.1.2
It would be useful to have a build of spark version 3.1.2 available.

1) Version 3.1.2 of spark matches the version that have deployed to
   the analytics cluster via the conda-analytics environment.

2) The spark-build container contains spark-3.1.2-yarn-shuffle.jar,
   which is not hosted on any external sources such as Maven Central.

3) Our conda-analytics environment install spark using a pyspark
   distribution, which also does not contain the spark-shuffler for
   yarn.

We need to provide spark-3.1.2-yarn-shuffle.jar to the Hadoop worker
nodes, so that they yarn resourcemanagers can load the required library.

Therefore, although we plan to keep upgrading the spark image in future,
this temporary downgrade will provide the necessary build artifact to
enable the upgrade of the yarn shuffler service.

SparkR has also been removed due to a build failure caused by:
https://issues.apache.org/jira/browse/SPARK-37843

Bug: T332765
Bug: T288804
Change-Id: I2dd99550c19469caebfad045a07956ea58d417c6