python-build: set date of source files in the wheels
When updating solely the dependencies of a python project, I would do a
diff of the wheels. bdist_wheel regenerates files with the current
running time which do not match the time from the previous build. That
leads to a bit of noise in the diff. An example for Zuul shows the
following files diffing solely by timestamp:

    zuul-2.5.2.dev30.data/scripts/zuul-clear-refs.py
    zuul-2.5.2.dev30.dist-info/AUTHORS
    zuul-2.5.2.dev30.dist-info/LICENSE
    zuul-2.5.2.dev30.dist-info/METADATA
    zuul-2.5.2.dev30.dist-info/WHEEL
    zuul-2.5.2.dev30.dist-info/dependency_links.txt
    zuul-2.5.2.dev30.dist-info/entry_points.txt
    zuul-2.5.2.dev30.dist-info/pbr.json
    zuul-2.5.2.dev30.dist-info/top_level.txt
    zuul-2.5.2.dev30.dist-info/RECORD

We can ask wheels to set the source file date by setting
SOURCE_DATE_EPOCH environment variable. This change set it to the Unix
epoch of the HEAD commiter date.

Bug: T259611
Bug: T342346
Change-Id: I90cceab6b3f42d4a9660c795e8acf4285037447a