python-build: provide a python2 Bullseye image
Our legacy Zuul application is hosted on a Buster based server which has
to migrate to Bullseye. We have determinated we can't reasonably port
Zuul to python3 and would use the unsupported python 2 provided by
Bullseye.

There are some shortfalls though, there is no python 2 based modules
available anymore. Notably `pip` is missing which we need to install
`wheel` and the Zuul dependencies. This downloads and install pip
version 20.3.4 which is the latest still supporting Python 2.7.

Bug: T342346
Change-Id: I8b118e94724dc625ab43bc519939580cea5666ec