Add temporary buster-based PHP7.4 icu67 images
Why:

Before we are able to migrate to PHP 8.1, we need to migrate first to
libicu 67. One plan to make this happen is sketched out in T329491 and
requires running shellbox with a PHP7.4 compiled against libicu67

What:

Add 2 images, almost identical to our current php7.4 images, with an
extra -icu67 suffix. This allows to proceed with the sketched out plan,
without impacting day to day operations/deployments of anything else,
selectively enabling shellbox to move to icu67

Bug: T329491
Bug: T345561
Change-Id: If4b584dec468267e1ddf73a6faf5b929c4ef0109