Refactor spark support to build multiple minor versions
We need to be able to run multiple versions of spark in parallel. This
includes the user facing binaries, the yarn shuffler service, and
potentially the kubernetes operator.

To-date we have only had a single version of the spark image pipelines
in the production-images repository and multiple versions have been
supported by simple, linear version changes.

This CR adds support for three minor versions (3.1, 3.3, and 3.4.
Currently our production pipelines are all on spark version 3.1 but
development work uses versions 3.3 and 3.4.

Versions 3.3 and 3.4 are built with support for R, whereas 3.1 is not.

Bug: T344910
Change-Id: I0fc43ae7b8d08a00124a58568e6970d8b76f8f56