python-build: add lsb-release, make and virtualenv
The deployment repositories have makefiles to build wheels and to deploy
on a target host. For local developement it is convenenient to have make
to test building the wheels and virtualenv to test their deployments.

I can then test the deployment of an applications of its wheel with
something such as:

  docker run --rm -it \
   -v .:/deploy \
   --entrypoint=make \
   <some image>
   --makefile=/deploy/src/Makefile.deploy

And then verify the application works using /deploy/venv/bin/python

Bug: T342346
Change-Id: I60da55e8c19fb70786c35f01e8e247df1bb073b7