kserve: upgrade all images to Bookworm
Since the build image runs on Bookworm (Golang 1.21), the resulting
binaries are built on it and they bind to a certain Glib version.

Bump everything to Bookworm since we are still running Buster,
and also add CGO_ENABLED=0 to add a safety net for future upgrades.

Refactor also the storage-initializer's image (running Python 3)
since on Bookworm the pip standards has been changed (namely
a venv is preferred).

Bug: T337213
Change-Id: Iaedac4c7f81fba59f17d705ee5eb3248f0e252a9