Create the /usr/share/binfmts directory to fix JRE error
We have experience build failures with our openjd-8-jre images due to an
upstream error in the postinst script. There is a patch for that
package, but it has not yet been released.

In order to work around the issue, we create a directory
/usr/share/binfmts which will allow the postinst script to remove it
without errors.

We will remove this workaround as soon as the upstream patch has been
released.

Bug: T358866
Change-Id: Ie12080e9f672700521402fad826ff9c4f0446fa9