Add new ceph container image
Use of cephadm requires a suitable container image to be available;
upstream's are based on centos. This is an attempt at a minimal set
for our intended purposes (i.e. no ganesha / mds etc.)

Based on upstream work in https://github.com/ceph/ceph/pull/54575/
which in turn is a move to simplify how upstream builds their
containers.

This is for the PoC ceph/reef multisite cluster, codenamed "apus".

Bug: T279621
Change-Id: I256acec29b5c50eff266690e15db066470038aa7