Rework the amd-pytorch22's image
Second version of the Pytorch base image for ML
workloads, that adds more complexity but it should
allow more flexibility for downstream users.

For example, anybody building their ML Docker image
via Blubber (using this image as base) should be able
to specify "torch" in their requirements.txt
(directly or as transitive dep) without worrying about
what's already present on the base image.

In order to do so, we need to mimic what Blubber does
and this sadly involves the creation of the "somebody"
user to be able to pip-install packages with correct
dependencies (so that downstream Blubber runs wouldn't
fail etc..). If the approach is accepted, another
companion puppet patch needs to be created to add
the "somebody" user in other places.

Bug: T360638
Change-Id: I3b9deca5c9d40bd2d8600ec4a2e727af56579517