amd-pytorch21: fix the ROCm version
The 5.7 version is not present in ROCm/Pytorch upstream,
and pip silently installed the Nvidia version, that
is almost the same except for the binaries that carries as
libraries. This in turn caused a bigger layer to be created,
that ended up to be rejected by the Docker registry due
to size limits (so the 5.7 images are not published).

This change fix the version to 5.6, and also removes the previous
entries since they shouldn't be pushed and/or checked
because totally wrong. The assumption is that we can remove them
since they haven't been published to the registry, but they
were only build locally by docker-pkg.

Bug: T362984
Change-Id: Ic03f088908b94b6f7c5a5c55e9705a3b0cc70a68