ceph: add mgr module for rgw (upstream packaging bug)
Upstream ship the ceph manager module for rgw (i.e. that enables the
deployment of more complex RGW setups from a spec file via cephadm),
but neglected to include it in the Debian packages.

This packaging mistake is being fixed upstream (
https://github.com/ceph/ceph/pull/57811 ) but in the mean time we need
this available for "apus".

The two python files are taken from the tip of the upstream reef
branch, commit fd81d117fbe

Change-Id: I5910f2ab68c230b513c55499e3a37bdff75822b2