Improve ability to override php-fpm configuration in kubernetes
By including the content of all files with "inc" extension contained in
a directory, we allow changing php-fpm settings, and in particular adding
environment variables to the container, via a kubernetes configmap instead
than needing an image rebuild every time.

In the process, we can now also move the mcrouter setting to the
multiversion-base image where it belongs.

Change-Id: I959d72588a173943dd6a669445ffef30fc87373f