Revert "Resurrect fluent-bit image"
This reverts commit a3d4a64daf02d7159246bafa3f223cf52c0bca39.

Reason for revert:

The functionality was apparently never used. After some discussions in
\#wikimedia-serviceops, and some inspections of the data flowing through
the pipeline powered by this component we collectively decided to turn
off that part of the infrastructure. Since that pipeline is the only
user of this component, we are going to revert the image creation and
delete the image.

Bug: T251812
Change-Id: Id01dc30eaa5390896e56b3591e690027624c7ab1