php-fpm: make /healthz smarter
In shellbox for videos, we want to have a single worker doing
videoscaling in a pod. If we set the total number of php-fpm workers
to 1, though, then all metrics requests and readiness probes will
hang until the videoscaling worker is done.

As a solution, we allow calls to /healthz to declare a minumum number
of workers that should be available to route traffic to the pod.

In the case of videoscaling shellboxes, we will set this to 1, and have
two workers in the php-fpm pool.

Change-Id: Ic08b3b950b603a137a004cb9f2ac196097aa3a70