php8.1-cli: initial release of 8.1-based image
Cloned from the analogous 7.4-based image.

The only notable difference is that php-geoip is not included, as it's
not supported on PHP 8.x.

Otherwise, this simply replaces 7.4 with 8.1, switches to a bullseye
base image, and resets the changelog. The same is true with the next two
patches (-fpm and -fpm-multiversion-base).

Smoke tested by building locally and running the entrypoint (no errors
or warnings).

Bug: T372602
Change-Id: I93ad4749a2a0385871b74687a2845cfcece11d4b