httpd: introduce -bookworm track and cascade
Introduces the "bookworm" track of httpd as httpd-bookworm, and cascades
through httpd-fcgi-bookworm and mediawiki-httpd-bookworm.

This is a transitional configuration for migration to bookworm-based
images, offering both the existing and bookworm tracks simultaneously.

Since there's no difference in apache2 version (2.4.62) between bullseye
and bookworm, we might as well jump right to bookworm.

Points of note:
* Only the Dockerfile.template files are duplicated, in order to pick up
  the relevant base image changes, while other resources are symlinked.
* Image version numbering is *reset* in httpd-bookworm and
  httpd-fcgi-bookworm in order to match the underlying apache2 package
  version.
* Image version numbering is *retained* (minus automated weekly rebuild
  timestamp) in mediawiki-httpd-bookworm for clarity, since its
  versioning is not tied to the inherited apache2 version.

Bug: T378128
Change-Id: I51382ebd49f591ee8cb2aad132ca1349127778fe