Update maintainership for all Java, Flink, ad Spark related container images
The Data Platform Engineering group and the Data Platform SRE team are
responsible for the maintenance of the following images collections in
this repository.

 * flink and its kubernetes operator
 * spark and its kubernetes operator

The SRE Infrastructure Foudations team is responsible for the OpenJDK
related images.

This commit updates the maintainership information in the control files
and bumps the version numbers in the changelog files accordingly.

Bug: T373534
Change-Id: I1b8e056d74dde4b7bcd73923e255f4337512f98f