Add the mapnik image
The Kartoterian service is being migrated to K8s, and as part of
its process to properly Dockerize the service image we had to
build upstream mapnik from scratch. The current blubber config
does it every time a change is submitted, not ideal if we want
to encourage devs to work on Kartotherian.

This change adds a base image for Kartotherian that builds
mapnik from source, to be used directly in Kartotherian's
blubber config.

The biggest change from the Kartotherian's blubber config is
that we don't need to use /opt anymore, but we can rely
to more standard /usr/local locations.

Note: This is an image that Kartotherian's Blubber config
will use while building stuff like node-mapnik and other
npm-install-related artifacts. The final Docker image
will use runtime dependencies and it will only copy
artifacts to /usr/local/{bin,include,lib}.

Note 2: We tried to package libmapnik-dev in a Debian
package, but we ended up in a rabbit hole and we decided
to use this solution instead. More info in:
https://phabricator.wikimedia.org/T327396#10375715

Bug: T327396
Change-Id: Ic8b441778a9164ce99fb5828a442b4bd35e600e5