Provide a base image for Rust, based on Bookworm using 'rustc-web' now at 1.78
This is the parallel Debian build used for Chromium and Firefox, and is
rather ahead of the default 1.63 one for Bookworm.

Per agreement with Alexandros, this is named "abstractwiki-rust-web" to make it
very clear that this image is only for use by the Abstract Wiki team and is not
provided for other teams.

Bug: T380807
Change-Id: Ic6b3091867d9280de9cada40da3b4edc82add217