mapnik: fix paths for mapnik directories
The /srv/{include, lib, bin) directories, created and populated by mapnik's build
process, should be copied over without the * expansion, because it
doesn't respect the original path. For example, the "mapnik" subdir
is not copied over, ending up with issues when running live traffic.

Rework the current logic to fix the current mapnik errors that
we are seeing in k8s staging.

Bug: T384285
Change-Id: I0e7864b337fb05806fee50aed3b8118aed310f02