knative: fix patch command and backport for patches for PSS migration
The first big thing to change is the patch command, that in the previous
way didn't clearly work but it went unnoticed (sigh, my bad).

While reviewing the patches to backport, I ended up also backporting
https://github.com/knative/serving/pull/13398 and
https://github.com/knative/serving/pull/13395. They both should be
needed to allow the "secure-pod-defaults" setting to be applied,
namely the knative controller will force automatically the right
securityContext options to pass the PSS restricted profile validations.

Bug: T369493
Change-Id: Ife2e3b95b084a5a3c9962a83a7ed5f9d2e84abf1