python3: add python3-venv to devel image
Part of virtualenv were upstreamed to Python3 standard library as the
`venv` module. It is lighter than virtualenv and can be preferred. To
install pip in the environment, `venv` uses `ensurepip` which is
packaged by Debian in `python3.9-venv` (which is confusing by itself).
Without it, `pytho3 -m venv` fails to provide pip and ask to install
`python3-venv`.

Change-Id: I22c013845fed6d1e102faf78724117e037342b83