httpd: Rebase on bookworm and cascade
This rebases the httpd image stack on debian bookworm, which will allow
us to stop building the bookworm-specific variants.

That will happen in a later patch once all previously-migrated services
are moved back to these images.

Also updates control files to consistently use a team-level Maintainer
alias, similar to what we've done in the bookworm-specific variants.

Note that weekly builds of this image stack are currently failing, due
the recent archival of buster on debian mirrors. After this change, they
work again.

Bug: T378128
Change-Id: I7c5c7a128e1733fb37a9e0e4ae147834d552d01f