httpd-fcgi: add missing image build dependency
Previously, use of the golang1.19 base image as a builder for glogger in
Dockerfile.template was missing from control.

During the weekly rebuild, this works only if docker-pkg happens to
finish building said image before reaching httpd-fcgi, otherwise the
latter will fail (since the golang1.19 image does not yet exist at the
tag referenced in the rendered Dockerfile).

This appears to be happening now and has resulted in weekly rebuild
failures for the last two weeks.

While I'm not sure what might have changed, my suspicion is that a
dependency on golang1.19 was removed from another image in the repo,
which previously (setting aside vagaries of os.walk traversal order)
forced the image earlier in the build chain.

Change-Id: Iab2d9c67e1717252d499d44063909fda3c44347c