python: Include virtualenv packages in python base image
Attempting to use `pip` to install packages using the `python` base
image will result in an error:

```
error: externally-managed-environment
× This environment is externally managed
╰─> To install Python packages system-wide, try apt install
    python3-xyz, where xyz is the package you are trying to
    install.

    If you wish to install a non-Debian-packaged Python package,
    create a virtual environment using python3 -m venv path/to/venv.
    Then use path/to/venv/bin/python and path/to/venv/bin/pip. Make
    sure you have python3-full installed.

    If you wish to install a non-Debian packaged Python application,
    it may be easiest to use pipx install xyz, which will manage a
    virtual environment for you. Make sure you have pipx installed.

    See /usr/share/doc/python3.11/README.venv for more information.
note: If you believe this is a mistake, please contact your Python
installation or OS distribution provider. You can override this, at the
risk of breaking your Python installation or OS, by passing
--break-system-packages.
hint: See PEP 668 for the detailed specification.
```

The right thing to do is to use `python3 -m venv <dir>` to set up a
virtualenv, but previously this base image did not include the
`python3-venv` package.  Now it does.

Also: Trimmed the list of packages in `python-devel` to just
`python3-dev`.

Change-Id: I805498743830f4607511051e3638748af8f2181b