php: remove deprecated ${} string interpolation
Removes the one instance of ${}-style string interpolation present in
the metrics-handler code shipped in /srv/monitoring.

This was deprecated in PHP 8.2 [0][1] and thus results in warnings in
mw-debug/next, which was recently upgraded to 8.3.

In this specific usage, it does not appear this style of interpolation
is structurally or functionally necessary, and should thus be safe to
remove.

Note: This does *not* bump the 7.4 changelogs, as the images are unused
and no longer build (will be deleted in T399729).

[0] https://www.php.net/manual/en/migration82.deprecated.php
[1] https://wiki.php.net/rfc/deprecate_dollar_brace_string_interpolation

Bug: T402424
Change-Id: I125e31d748968a6db0a2ba0010a86abb2af3df46