envoy-future: Update to v1.32.12
Also, set an internal_address_config in envoy-basic-config.yaml. This
cleans up a startup warning in test.sh (since the 1.29 upgrade) but
doesn't change anything in production and doesn't even need to be rolled
out, really: all our prod Envoy installations have their own config.

As a result, slightly unorthodox but not problematic: image/envoy-future
brings in that config via symlink from images/envoy, and we're editing
it at the source but only bumping the envoy-future changelog. The envoy
image will harmlessly pick it up with the next weekly rebuild.

Bug: T405808
Bug: T404036
Change-Id: I30667baceba5ae87e2adba3f43919e1f6a6a6964